Changed java update to 7

+++ java.sh
@@ -3,21 +3,21 @@
 
 . ./common.sh
 
-JDK_DIR=jdk1.7.0_05
+JDK_DIR=jdk1.7.0_07
 
 install_java() {
 #Install java, assumes download present in ./tmp
-JAVA_PACKAGE=jdk-7u5-linux-x64.tar.gz
+JAVA_PACKAGE=jdk-7u7-linux-x64.tar.gz
 
-#if [ ! -f ./tmp/$JAVA_PACKAGE ]
-#then
-#  echo Please download java package to tmp folder!
-#  exit
-#fi
-#sudo mkdir -p /usr/lib/jvm/
-#sudo mv ./tmp/$JAVA_PACKAGE /usr/lib/jvm/
-#cd /usr/lib/jvm/
-#sudo tar zxvf $JAVA_PACKAGE
+if [ ! -f ./tmp/$JAVA_PACKAGE ]
+then
+  echo Please download java package to tmp folder!
+  exit
+fi
+sudo mkdir -p /usr/lib/jvm/
+sudo mv ./tmp/$JAVA_PACKAGE /usr/lib/jvm/
+cd /usr/lib/jvm/
+sudo tar zxvf $JAVA_PACKAGE
 
 sudo update-alternatives --install "/usr/bin/java" "java" "/usr/lib/jvm/$JDK_DIR/bin/java" 1062
 sudo update-alternatives --install "/usr/bin/javac" "javac" "/usr/lib/jvm/$JDK_DIR/bin/javac" 1062
